Found what look like a couple of bugs:
1) In the read_write example, in a write to descriptor, the buffer is never written with the value to write.

2) In /bluez/gattlib_read_write.c the sizeof the pointer is used rather than the length of the buffer when writing the descriptor

I think there may still be problems in the read write example with regard to the  type of the value you wish to write.
In my case it is a uint8_t. I used the example here
https://github.com/sandeepmistry/arduino-BLEPeripheral/blob/master/examples/led/led.ino

Anyway with the version in my branch I can now write the value correctly, but probbaly needs work to be generic.